BeingDataLoader uses event stream instead of signal
Many UI widgets are signals (behaviors that hold an initial value),
and in a typical beings UI, the initial values are  bogus (empty
strings etc.) that gets replaced with the being data loaded from
files as soon as its available. We do not want these initial bogus
events, therefore even streams and not signals.